LayoutTests/imported/w3c/web-platform-tests/selection/textcontrols/selectionchange-on-shadow-dom-expected.txt


FAIL selectionchange event fired on a shadow dom bubble to the document assert_equals: expected 1 but got 0
